Make SDP_Table an array of int16_t.
Move some functions into an anonymous namespace.

R=tsepez@chromium.org

Review URL: https://codereview.chromium.org/1431253002 .

+++ core/src/fxge/dib/dib_int.h
@@ -7,12 +7,14 @@
 #ifndef CORE_SRC_FXGE_DIB_DIB_INT_H_
 #define CORE_SRC_FXGE_DIB_DIB_INT_H_
 
+#include <stdint.h>
+
 #include "core/include/fxcrt/fx_coordinates.h"
 #include "core/include/fxge/fx_dib.h"
 
 class IFX_ScanlineComposer;
 
-extern const int SDP_Table[513];
+extern const int16_t SDP_Table[513];
 
 class CPDF_FixedMatrix {
  public:
